Revert "scripts/builddeb: add support for source only generation"
This reverts commit ae29d0598c2095114b47e927333f1bd802754d98.
Intend of this patch was to  prepare  tarball with sources
from odp.git. But to do that ./configure stage should be passed
to create 'dist' target in Makefile. But ./configure checks
for various discrepancies and traps if they do not match. Patch
reverted due to it does not solve original problem.

Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>